警告なおした

+++ FrameWork/MyEffect/MyEffect.cpp
@@ -28,6 +28,6 @@ void MyEffect::Update(DX::StepTimer timer)
 	//	Initialize(m_startLife, m_startPosition, m_startVelocity);
 
 	//}
-	m_life -= m_timer.GetElapsedSeconds();
+	m_life -= static_cast<float>(m_timer.GetElapsedSeconds());
 }
